% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute.R
\name{bootstrap_CI_pearson}
\alias{bootstrap_CI_pearson}
\title{Confidence interval for the correlation between residuals and treatment}
\usage{
bootstrap_CI_pearson(prediction, outcome, treatment, iters, alpha = 0.05,
  side = "both")
}
\arguments{
\item{prediction}{Vector of predicted outcomes}

\item{treatment}{Vector of treatments}

\item{iters}{The number of Monte Carlo iterations (default 1000)}

\item{alpha}{Significance level}

\item{side}{Type of interval, either "both", "upper", or "lower". Default is "both".}
}
\value{
a confidence interval (vector)
}
\description{
Bootstrap to get a \eqn{1-\alpha} confidence interval for the Pearson correlation between residuals and treatment
}

