% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute.R
\name{permu_CI_mean}
\alias{permu_CI_mean}
\title{Confidence interval for the difference in means within groups}
\usage{
permu_CI_mean(groups, prediction, response, treatment, side = "both",
  alpha = 0.05, iters = 1000)
}
\arguments{
\item{groups}{List output from Matches, containing matched pairs or groups}

\item{prediction}{Vector of predicted outcomes}

\item{response}{Vector of responses}

\item{treatment}{Vector of treatments}

\item{side}{Type of interval, either "both", "upper", or "lower". Default is "both".}

\item{alpha}{Significance level}

\item{iters}{The number of Monte Carlo iterations (default 1000)}

\item{precision}{Rate at which to iteratively increment the confidence bounds, relative to the estimated difference in means. The value should be between 0 and 1. Smaller precision results in a slower run time but more precise confidence bounds (default is 0.02)}
}
\value{
a vector of differences
}
\description{
Invert the stratified permutation test to get a \eqn{1-\alpha} confidence interval for the difference in mean residuals
}

