% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permu_pearson}
\alias{permu_pearson}
\title{Test the correlation between residuals and treatment}
\usage{
permu_pearson(prediction, response, treatment, iters = 1000, shift = 0)
}
\arguments{
\item{prediction}{Vector of predicted outcomes}

\item{response}{Vector of responses}

\item{treatment}{Vector of treatments}

\item{iters}{The number of Monte Carlo iterations (default 1000)
#' @param shift Null shift away from 0 (a scalar constant) in correlation (default 0)}
}
\value{
a list containing the results: attributes estimate (the estimated difference), distr (simulated permutation distribution), and pvalue (p-value for the test)
}
\description{
Carry out a permutation test for the Pearson correlation between residuals and treatment.
}

