% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permu_test_mean_slow}
\alias{permu_test_mean_slow}
\title{Test for the difference in means within groups}
\usage{
permu_test_mean_slow(strata, prediction, treatment, response, iters = 1000,
  shift = 0)
}
\arguments{
\item{strata}{List output from Matches, containing matched pairs or groups}

\item{prediction}{Vector of predicted outcomes}

\item{treatment}{Vector of treatments}

\item{response}{Vector of responses}

\item{iters}{The number of Monte Carlo iterations (default 1000)}

\item{shift}{Null shift (a scalar constant) for the treatment group (default 0)}
}
\value{
a list containing the results: attributes diff_means (the estimated difference), perm_distribution (simulated permutation distribution), and pvalue (p-value for the test)
}
\description{
Carry out stratified permutation test for difference in mean residuals between the treatment and control groups.
}

