% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute_within_groups}
\alias{permute_within_groups}
\title{Permute within groups}
\usage{
permute_within_groups(groups)
}
\arguments{
\item{groups}{List output from Matches, containing matched pairs or groups}
}
\value{
a list of the same structure as the input, with treatment assignments permuted.
}
\description{
Permute treatment assignment within groups of matched individuals
}

