% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{Strata}
\alias{Strata}
\title{Stratify on model predictions.}
\usage{
Strata(treatment, prediction, strata = 5)
}
\arguments{
\item{treatment}{Vector of treatment values}

\item{prediction}{Vector of predicted outcomes}

\item{strata}{Number of strata desired. Default is 5.}
}
\value{
a list. Each entry is a group of matched individuals with their treatments.
}
\description{
Note, we need sufficient distribution of the treatment values within each bin. User should check this.
}

