% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute.R
\name{Matches}
\alias{Matches}
\title{Match on model predictions.}
\usage{
Matches(treatment, prediction)
}
\arguments{
\item{treatment}{Vector of treatment values}

\item{prediction}{Vector of predicted outcomes}
}
\value{
a list. Each entry is a group of matched individuals with their treatments.
}
\description{
At this point, Matches only supports one-to-one matching of treatment and control groups.
}

