% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute.R
\name{permu_CI_pearson}
\alias{permu_CI_pearson}
\title{Confidence interval for the correlation between residuals and treatment}
\usage{
permu_CI_pearson(prediction, response, treatment, iters, alpha = 0.05,
  side = "both", verbosity = FALSE)
}
\arguments{
\item{prediction}{Vector of predicted outcomes}

\item{response}{Vector of responses}

\item{treatment}{Vector of treatments}

\item{iters}{The number of Monte Carlo iterations (default 1000)}

\item{alpha}{Significance level}

\item{side}{Type of interval, either "both", "upper", or "lower". Default is "both".}

\item{verbose}{Verbosity switch - print the p-value and confidence interval endpoint at each step? (Default FALSE)}
}
\value{
a confidence interval (vector)
}
\description{
Invert the permutation test to get a \eqn{1-\alpha} confidence interval for the Pearson correlation between residuals and treatment
}

