% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permu_pearson}
\alias{permu_pearson}
\title{Test the correlation between residuals and treatment}
\usage{
permu_pearson(prediction, response, treatment, iters = 1000)
}
\arguments{
\item{prediction}{Vector of predicted outcomes}

\item{response}{Vector of responses}

\item{treatment}{Vector of treatments}

\item{iters}{The number of Monte Carlo iterations (default 1000)}
}
\value{
a list containing the results: attributes estimate (the estimated difference), distr (simulated permutation distribution), and pvalue (p-value for the test)
}
\description{
Carry out a permutation test for the Pearson correlation between residuals and treatment.
}

