% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permute.R
\name{within_group_mean}
\alias{within_group_mean}
\title{Difference in means within groups}
\usage{
within_group_mean(groups, prediction, response, shift = 0)
}
\arguments{
\item{groups}{List output from Matches, containing matched pairs or groups}

\item{prediction}{Vector of predicted outcomes}

\item{response}{Vector of responses}

\item{shift}{Null shift (a scalar constant) for the treatment group (default 0)}
}
\value{
a vector of differences
}
\description{
Compute difference in mean residual between treated and control within each group
}

